Separate swapchain + split to dedicated files.
